+++ src/Exception/Exception.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TomPHP\Transform\Exception;
+
+/**
+ * Used to identify exceptions from this package.
+ */
+interface Exception
+{
+}